[ADD] MacOS compatibility
